meu-chatbot/requirements.txt
python-telegram-bot==13.7
python-dotenv==0.19.0
requests==2.26.0
google.generativeai